requirements.txt
# requirements.txt
pyMySQL
mysql-connector-python
pydantic
fastapi
uvicorn
sqlalchemy
alembic
python-magic
pytest

